Parallel State tracking
