+++ common/http/src/lib.rs
@@ -5,7 +5,6 @@ pub mod batch;
 pub mod client;
 pub mod limit;
 pub mod metrics_endpoint;
-pub mod offsets;
 pub mod retry;
 
 pub mod types {
