+++ common/state/Cargo.toml
@@ -11,11 +11,11 @@ derivative = "2.2"
 futures = "0.3"
 rocksdb = "0.15"
 async-channel = "1.6"
-async-lock = "2.4.0"
 thiserror = "1.0"
-bytes = "1"
 log = "0.4"
+serde = { version = "1", features = ["derive", "rc"] }
 slotmap = "1"
+vec-collections = "0.3"
 
 [dev-dependencies]
 env_logger = "0.8"
